Improve cropping operaions

+++ vcam.h
@@ -1,6 +1,8 @@
 #ifndef VCAM_H
 #define VCAM_H
 
+#include <asm/types.h>
+
 #define VCAM_IOCTL_CREATE_DEVICE 0x111
 #define VCAM_IOCTL_DESTROY_DEVICE 0x222
 #define VCAM_IOCTL_GET_DEVICE 0x333
@@ -11,7 +13,7 @@ typedef enum { VCAM_PIXFMT_RGB24 = 0x01, VCAM_PIXFMT_YUYV = 0x02 } pixfmt_t;
 
 struct vcam_device_spec {
     unsigned int idx;
-    unsigned short width, height;
+    __u32 width, height;
     pixfmt_t pix_fmt;
     char video_node[64];
     char fb_node[64];
